Fix cpplint warning on nlohmann serializer
